+++ dart/lib/src/utils/stacktrace_utils.dart
@@ -8,3 +8,11 @@ import 'package:meta/meta.dart';
 // https://github.com/getsentry/sentry/blob/master/src/sentry/utils/sdk_crashes/README.rst
 @internal
 StackTrace getCurrentStackTrace() => StackTrace.current;
+
+/// Regex that matches an “abs …” stack-frame line emitted by split-debug-info builds.
+@internal
+final absRegex = RegExp(r'^\s*#[0-9]+ +abs +([A-Fa-f0-9]+)');
+
+/// Regex that matches the header of a stacktrace emitted by split-debug-info builds.
+@internal
+final buildIdRegex = RegExp(r"build_id[:=] *'([A-Fa-f0-9]+)'");