+++ flutter/lib/src/sentry_flutter.dart
@@ -23,6 +23,7 @@ import 'integrations/frames_tracking_integration.dart';
 import 'integrations/integrations.dart';
 import 'integrations/native_app_start_handler.dart';
 import 'integrations/screenshot_integration.dart';
+import 'integrations/generic_app_start_integration.dart';
 import 'integrations/web_session_integration.dart';
 import 'native/factory.dart';
 import 'native/native_scope_observer.dart';
@@ -196,12 +197,14 @@ mixin SentryFlutter {
           integrations.add(LoadContextsIntegration(native));
         }
         integrations.add(FramesTrackingIntegration(native));
-        integrations.add(
-          NativeAppStartIntegration(
-            DefaultFrameCallbackHandler(),
-            NativeAppStartHandler(native),
-          ),
-        );
+        if (platform.isIOS || platform.isAndroid || platform.isMacOS) {
+          integrations.add(
+            NativeAppStartIntegration(
+              DefaultFrameCallbackHandler(),
+              NativeAppStartHandler(native),
+            ),
+          );
+        }
         integrations.add(ReplayIntegration(native));
       } else {
         // Updating sessions manually is only relevant for web
@@ -212,6 +215,10 @@ mixin SentryFlutter {
         integrations.add(WebSessionIntegration(native));
       }
       options.enableDartSymbolication = false;
+    }
+
+    if (platform.isWeb || platform.isLinux || platform.isWindows) {
+      integrations.add(GenericAppStartIntegration());
     }
 
     if (options.isScreenshotSupported) {
@@ -296,6 +303,8 @@ mixin SentryFlutter {
     }
     final transactionId = options.timeToDisplayTracker.transactionId;
     if (transactionId == null) {
+      hub.options.log(SentryLevel.error,
+          'Could not process TTFD for screen ${SentryNavigatorObserver.currentRouteName} - transactionId should not be null');
       return null;
     }
     return SentryDisplay(transactionId, hub: hub);