+++ flutter/test/sentry_flutter_test.dart
@@ -11,6 +11,7 @@ import 'package:sentry_flutter/src/flutter_exception_type_identifier.dart';
 import 'package:sentry_flutter/src/integrations/connectivity/connectivity_integration.dart';
 import 'package:sentry_flutter/src/integrations/integrations.dart';
 import 'package:sentry_flutter/src/integrations/screenshot_integration.dart';
+import 'package:sentry_flutter/src/integrations/generic_app_start_integration.dart';
 import 'package:sentry_flutter/src/integrations/web_session_integration.dart';
 import 'package:sentry_flutter/src/profiling.dart';
 import 'package:sentry_flutter/src/renderer/renderer.dart';
@@ -37,6 +38,10 @@ final platformAgnosticIntegrations = [
 final webIntegrations = [
   ConnectivityIntegration,
   WebSessionIntegration,
+];
+
+final linuxWindowsAndWebIntegrations = [
+  GenericAppStartIntegration,
 ];
 
 final nonWebIntegrations = [
@@ -258,6 +263,7 @@ void main() {
         shouldHaveIntegrations: [
           ...platformAgnosticIntegrations,
           ...nonWebIntegrations,
+          ...linuxWindowsAndWebIntegrations,
         ],
         shouldNotHaveIntegrations: [
           ...androidIntegrations,
@@ -306,6 +312,7 @@ void main() {
         shouldHaveIntegrations: [
           ...platformAgnosticIntegrations,
           ...nonWebIntegrations,
+          ...linuxWindowsAndWebIntegrations,
         ],
         shouldNotHaveIntegrations: [
           ...androidIntegrations,
@@ -355,6 +362,7 @@ void main() {
         shouldHaveIntegrations: [
           ...platformAgnosticIntegrations,
           ...webIntegrations,
+          ...linuxWindowsAndWebIntegrations,
         ],
         shouldNotHaveIntegrations: [
           ...androidIntegrations,
@@ -425,6 +433,7 @@ void main() {
         shouldHaveIntegrations: [
           ...platformAgnosticIntegrations,
           ...webIntegrations,
+          ...linuxWindowsAndWebIntegrations,
         ],
         shouldNotHaveIntegrations: [
           ...androidIntegrations,
@@ -469,6 +478,7 @@ void main() {
         shouldHaveIntegrations: [
           ...platformAgnosticIntegrations,
           ...webIntegrations,
+          ...linuxWindowsAndWebIntegrations,
         ],
         shouldNotHaveIntegrations: [
           ...androidIntegrations,
@@ -513,6 +523,7 @@ void main() {
         shouldHaveIntegrations: [
           ...platformAgnosticIntegrations,
           ...webIntegrations,
+          ...linuxWindowsAndWebIntegrations,
         ],
         shouldNotHaveIntegrations: [
           ...androidIntegrations,